requirements.txt
requests==2.21.0
xlrd==1.1.0
pymongo==3.9.0
Werkzeug==0.15.6
PyJWT==1.7.1
flask-restplus==0.13.0
uWSGI==2.0.18
lxml
flask-cors
